Finished refactoring column and Element class
